feat: similar papers endpoint
