+++ backend-rs/src/routing/mod.rs
@@ -29,6 +29,7 @@ pub fn get_router(env_vars: &EnvVars, db: Database) -> axum::Router {
         .route("/profile", axum::routing::get(handlers::profile))
         .route("/edit", axum::routing::post(handlers::edit))
         .route("/delete", axum::routing::post(handlers::delete))
+        .route("/similar", axum::routing::post(handlers::similar))
         .route_layer(axum::middleware::from_fn_with_state(
             state.clone(),
             middleware::verify_jwt_middleware,
